automated-research-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
cerebras-cloud-sdk==1.0.0
langchain-cerebras>=0.1.0
langsmith>=0.1.0
fastapi-cors==0.0.6